C implementation of dirty vanity with an additional option
C implementation of dirty vanity with additional option available. In the presentation and Github repo, Eliran created a prototype implementation where a user can input a PID as an argument, and the "dirty vanity" will inject and run a message box in the forked process of the original provided process. At the end of author's presentation, he mentioned that the dirty vanity technique could be expanded to more variations, such as NtCreateProcess[Ex] + Execute primitive. I followed the idea and implemented Dirty Vanity with an additional option. If a user does not provide a PID of a running process, Dirty Vanity will open a calculator and use the calculator's PID to allocate space and inject the shellcode. Then, Vanity will fork a calculator process and execute the shellcode.